Add Mongoid6, refactor code, drop support of old versions
* Add support of Mongoid6
* Add new Rails / Ruby to Travis build matrix
* Move gem to depend on Doorkeeper 4
* Remove old Mongoid versions
