--- gemfiles/Gemfile.mongoid2.rb
@@ -1,5 +1,0 @@
-gemfile = File.expand_path("../Gemfile.common.rb", __FILE__)
-instance_eval IO.read(gemfile), gemfile
-
-gem 'mongoid', '~> 2'
-gem 'bson_ext'
