--- gemfiles/Gemfile.mongoid3.rb
@@ -1,4 +1,0 @@
-gemfile = File.expand_path("../Gemfile.common.rb", __FILE__)
-instance_eval IO.read(gemfile), gemfile
-
-gem 'mongoid', '~> 3'
