+++ gemfiles/Gemfile.mongoid6.rb
@@ -1,0 +1,4 @@
+gemfile = File.expand_path("../Gemfile.common.rb", __FILE__)
+instance_eval IO.read(gemfile), gemfile
+
+gem 'mongoid', '~> 6'
