+++ lib/doorkeeper-mongodb.rb
@@ -2,10 +2,9 @@ require 'doorkeeper-mongodb/version'
 
 require 'doorkeeper'
 
-require 'doorkeeper/orm/mongoid2'
-require 'doorkeeper/orm/mongoid3'
 require 'doorkeeper/orm/mongoid4'
 require 'doorkeeper/orm/mongoid5'
+require 'doorkeeper/orm/mongoid6'
 require 'doorkeeper/orm/mongo_mapper'
 
 module DoorkeeperMongodb
