+++ lib/doorkeeper-mongodb/version.rb
@@ -1,3 +1,3 @@
 module DoorkeeperMongodb
-  VERSION = "3.0.0"
+  VERSION = '3.0.0'.freeze
 end
