+++ lib/doorkeeper/orm/mongo_mapper/application.rb
@@ -19,9 +19,10 @@ module Doorkeeper
 
     def self.authorized_for(resource_owner)
       ids = AccessToken.where(
-          resource_owner_id: resource_owner.id,
-          revoked_at: nil
-        ).map(&:application_id)
+        resource_owner_id: resource_owner.id,
+        revoked_at: nil
+      ).map(&:application_id)
+
       find(ids)
     end
