+++ lib/doorkeeper/orm/mongoid6/concerns/scopes.rb
@@ -1,0 +1,17 @@
+module Doorkeeper
+  module Models
+    module Mongoid6
+      module Scopes
+        extend ActiveSupport::Concern
+
+        included do
+          field :scopes, type: String
+        end
+
+        def scopes=(value)
+          write_attribute :scopes, value
+        end
+      end
+    end
+  end
+end
