--- spec/dummy/config/mongoid2.yml
@@ -1,9 +1,0 @@
-development:
-  database: doorkeeper-development
-  persist_in_safe_mode: true
-  autocreate_indexes: true
-
-test:
-  database: doorkeeper-mongoid2-test
-  persist_in_safe_mode: true
-  autocreate_indexes: true
