--- spec/dummy/config/mongoid3.yml
@@ -1,18 +1,0 @@
-development:
-  sessions:
-    default:
-      database: doorkeeper-mongoid3-development
-      hosts:
-        - localhost:27017
-      options:
-        consistency: :strong
-        safe: true
-test:
-  sessions:
-    default:
-      database: doorkeeper-mongoid3-test
-      hosts:
-        - localhost:27017
-      options:
-        consistency: :strong
-        safe: true
