--- spec/support/orm/mongoid2.rb
@@ -1,1 +1,0 @@
-require_relative "mongoid.rb"
