--- spec/support/orm/mongoid3.rb
@@ -1,1 +1,0 @@
-require_relative "mongoid.rb"
